requirements.txt
absl-py==0.14.1
appdirs==1.4.4
astunparse==1.6.3
black==21.5b1
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
filelock==3.3.0
flatbuffers==1.12
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.2
joblib==1.0.1
Keras-Preprocessing==1.1.2
Markdown==3.3.4
mypy-extensions==0.4.3
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pathspec==0.9.0
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyflakes==2.3.1
pyparsing==2.4.7
python-dotenv==0.17.1
regex==2021.9.30
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.46
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
tensorflow-hub==0.12.0
termcolor==1.1.0
tokenizers==0.10.3
toml==0.10.2
torch==1.9.1
tqdm==4.62.3
transformers==4.5.1
typing-extensions==3.7.4.3
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.12.1
